Re-enable integration tests for quarkus
