+++ platform/quarkus/src/test/java/org/trellisldp/ext/webapp/PgsqlLdpTest.java
@@ -14,24 +14,30 @@
 package org.trellisldp.ext.webapp;
 
 import static javax.ws.rs.client.ClientBuilder.newBuilder;
+import static org.junit.jupiter.api.condition.OS.WINDOWS;
 
 import io.quarkus.test.common.http.TestHTTPResource;
+import io.quarkus.test.junit.QuarkusTest;
 
 import java.net.URL;
 
 import javax.ws.rs.client.Client;
 
-import org.trellisldp.test.MementoTimeMapTests;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.condition.DisabledOnOs;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
+import org.trellisldp.test.AbstractApplicationLdpTests;
 
-abstract class AbstractMementoTimeMapTests implements MementoTimeMapTests {
+@DisabledOnOs(WINDOWS)
+@EnabledIfSystemProperty(named = "trellis.test.quarkus.external-pgsql", matches = "true")
+@QuarkusTest
+class PgsqlLdpTest extends AbstractApplicationLdpTests {
 
     private static final Client client = newBuilder().build();
 
     @TestHTTPResource
     URL url;
-
-    private String resource;
-    private String binary;
 
     @Override
     public Client getClient() {
@@ -43,23 +49,13 @@ abstract class AbstractMementoTimeMapTests implements MementoTimeMapTests {
         return url.toString();
     }
 
-    @Override
-    public void setBinaryLocation(final String location) {
-        binary = location;
+    @BeforeAll
+    static void setUp() throws Exception {
+        System.setProperty("quarkus.datasource.url", "jdbc:postgresql://localhost/trellis");
     }
 
-    @Override
-    public String getBinaryLocation() {
-        return binary;
-    }
-
-    @Override
-    public void setResourceLocation(final String location) {
-        resource = location;
-    }
-
-    @Override
-    public String getResourceLocation() {
-        return resource;
+    @AfterAll
+    static void tearDown() throws Exception {
+        System.clearProperty("quarkus.datasource.url");
     }
 }
